More background stuff, still developing

+++ background.py
@@ -1,4 +1,6 @@
 from flask import Blueprint, jsonify, render_template, request
+import os
+from subprocess import call
 import supervisor.xmlrpc
 import xmlrpclib
 
@@ -8,23 +10,38 @@ bg = xmlrpclib.ServerProxy('http://127.0.0.1', transport=supervisor.xmlrpc.Super
 editor = Blueprint('background', __name__, template_folder='templates')
 
 @editor.route('/editor/background/add/<scriptname>', methods=['GET', 'POST'])
-def add():
+def add(scriptname):
     # write config file in /etc/supervisor/conf.d
     conffile = '/etc/supervisor/conf.d/' + scriptname + '.conf'
-    with open(conffile, 'w') as f:
-        f.write('[program:' + scriptname + ']')
-        f.write('command=/var/www/public/background/' + scriptname + '.py')
-        f.write('autostart=true')
-        f.close()
+    scriptfile = '/var/www/public/background/' + scriptname + '.py'
+    if not os.path.exists(scriptfile):
+        with open(scriptfile, 'w') as s:
+            s.write('# This is an empty Python file created just for you.')
+            s.close()
+        os.chmod(scriptfile, 0744) # Mark file executable
+    if not os.path.exists(conffile):
+        with open(conffile, 'w') as f:
+            f.write('[program:' + scriptname + ']\n')
+            f.write('command=' + scriptfile + '\n')
+            f.write('autostart=true\n')
+            f.close()
+        call(["supervisorctl", "update"])
+        return jsonify(result=0, msg='Script ' + scriptname + ' added as a background process.')
+    else:
+        return jsonify(result=1, msg='The script ' + scriptname + ' is already configured, so no changes were made.')
 
 @editor.route('/editor/background/status', methods=['GET', 'POST'])
 def allstatus():
     d = bg.supervisor.getAllProcessInfo()
-    return jsonify(**d)
+    print d
+    names = dict(zip([script['name'] for script in d], [script['statename'] + ': ' + script['description'] for script in d]))
+    print names
+    return jsonify(**names)
 
 @editor.route('/editor/background/status/<scriptname>', methods=['GET', 'POST'])
 def status(scriptname):
     d = bg.supervisor.getProcessInfo(scriptname)
+    # Needs error checking
     return jsonify(**d)
 
 @editor.route('/editor/background/start/<scriptname>', methods=['GET', 'POST'])
